pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "x4-wallpaper-maker"
version = "0.1.0"
description = "Local-only desktop utility for CrossPoint-compatible Xteink X4 wallpapers."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "Pillow>=10.0.0,<12.0.0",
    "PySide6>=6.7.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-qt>=4.4.0,<5.0.0",
]

[project.scripts]
x4-wallpaper-maker = "x4_wallpaper_maker.app:main"

[tool.setuptools.packages.find]
include = ["x4_wallpaper_maker*"]

